[fix #32] 게이트웨이 swagger 설정 수정
