--- gateway/src/main/java/com/ticketPing/gateway/infrastructure/repository/RedisRepository.java
@@ -1,31 +1,0 @@
-package com.ticketPing.gateway.infrastructure.repository;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class RedisRepository {
-
-    private final RedisTemplate<String, Object> redisTemplate;
-    private final ObjectMapper objectMapper;
-
-    public Object getValue(String key) {
-        return redisTemplate.opsForValue().get(key);
-    }
-
-    public <T> T getValueAsClass(String key, Class<T> clazz) {
-        return objectMapper.convertValue(redisTemplate.opsForValue().get(key), clazz);
-    }
-
-    public Long getSortedSetSize(String key) {
-        return redisTemplate.opsForZSet().size(key);
-    }
-
-    public Boolean hasKey(String key) {
-        return redisTemplate.hasKey(key);
-    }
-
-}
